feat : Add image Column to table
- 이미지컬럼을 추가한다

issue: #7
