--- BE/src/main/java/com/team10/banchan/model/TopImage.java
@@ -1,24 +1,0 @@
-package com.team10.banchan.model;
-
-public class TopImage {
-
-    private final String alt;
-    private final String topImage;
-
-    TopImage(String alt, String topImage) {
-        this.alt = alt;
-        this.topImage = topImage;
-    }
-
-    public static TopImage of(String alt, String topImage) {
-        return new TopImage(alt, topImage);
-    }
-
-    public String getAlt() {
-        return alt;
-    }
-
-    public String getUrl() {
-        return topImage;
-    }
-}
